homework/model-engine: CLI for training, testing and saving/loading models
Problem:
With 2GB of memory on local laptop (1.2GB free, other stuff used by OS),
I often run into out of memory problems.
To get over it for the first part of fastai course, I've built a script
to save/load trained models to file.
With this, I can train big model, save to file, reload and play with it,
without fear of losing long-trained-model, if the process crashes.